grpc logging from relays
