+++ relay/server.go
@@ -5,10 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"strings"
 	"time"
 
-	"github.com/Layr-Labs/eigenda/api"
 	pb "github.com/Layr-Labs/eigenda/api/grpc/relay"
 	"github.com/Layr-Labs/eigenda/common/healthcheck"
 	"github.com/Layr-Labs/eigenda/common/pprof"
@@ -24,7 +22,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/keepalive"
-	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/reflection"
 )
 
@@ -186,383 +183,6 @@ func NewServer(
 	healthcheck.RegisterHealthServer(name, server.grpcServer)
 
 	return server, nil
-}
-
-// GetBlob retrieves a blob stored by the relay.
-func (s *Server) GetBlob(ctx context.Context, request *pb.GetBlobRequest) (*pb.GetBlobReply, error) {
-	start := time.Now()
-
-	if s.config.Timeouts.GetBlobTimeout > 0 {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, s.config.Timeouts.GetBlobTimeout)
-		defer cancel()
-	}
-
-	// Validate the request params before any further processing (as validation is cheaper)
-	key, err := v2.BytesToBlobKey(request.GetBlobKey())
-	if err != nil {
-		return nil, api.NewErrorInvalidArg(fmt.Sprintf("invalid blob key: %v", err))
-	}
-	s.logger.Debug("GetBlob request received", "key", key.Hex())
-
-	err = s.blobRateLimiter.BeginGetBlobOperation(time.Now())
-	if err != nil {
-		return nil, api.NewErrorResourceExhausted(fmt.Sprintf("rate limit exceeded: %v", err))
-	}
-	defer s.blobRateLimiter.FinishGetBlobOperation()
-
-	keys := []v2.BlobKey{key}
-	mMap, err := s.metadataProvider.GetMetadataForBlobs(ctx, keys)
-	if err != nil {
-		if strings.Contains(err.Error(), blobstore.ErrMetadataNotFound.Error()) {
-			// nolint:wrapcheck
-			return nil, api.NewErrorNotFound(
-				fmt.Sprintf("blob %s not found, check if blob exists and is assigned to this relay", key.Hex()))
-		}
-		// nolint:wrapcheck
-		return nil, api.NewErrorInternal(fmt.Sprintf("error fetching metadata for blob: %v", err))
-	}
-	metadata := mMap[v2.BlobKey(request.GetBlobKey())]
-	if metadata == nil {
-		return nil, api.NewErrorNotFound("blob not found")
-	}
-
-	finishedFetchingMetadata := time.Now()
-	s.metrics.ReportBlobMetadataLatency(finishedFetchingMetadata.Sub(start))
-
-	s.metrics.ReportBlobRequestedBandwidthUsage(int(metadata.blobSizeBytes))
-	err = s.blobRateLimiter.RequestGetBlobBandwidth(time.Now(), metadata.blobSizeBytes)
-	if err != nil {
-		return nil, api.NewErrorResourceExhausted(fmt.Sprintf("bandwidth limit exceeded: %v", err))
-	}
-
-	data, err := s.blobProvider.GetBlob(ctx, key)
-	if err != nil {
-		if strings.Contains(err.Error(), blobstore.ErrBlobNotFound.Error()) {
-			return nil, api.NewErrorNotFound(fmt.Sprintf("blob %s not found", key.Hex()))
-		} else {
-			s.logger.Errorf("error fetching blob %s: %v", key.Hex(), err)
-			return nil, api.NewErrorInternal(
-				fmt.Sprintf("relay encountered errors while attempting to fetch blob %s", key.Hex()))
-		}
-	}
-
-	s.metrics.ReportBlobBandwidthUsage(len(data))
-	s.metrics.ReportBlobDataLatency(time.Since(finishedFetchingMetadata))
-	s.metrics.ReportBlobLatency(time.Since(start))
-
-	reply := &pb.GetBlobReply{
-		Blob: data,
-	}
-	return reply, nil
-}
-
-func (s *Server) validateGetChunksRequest(request *pb.GetChunksRequest) error {
-	if request == nil {
-		return api.NewErrorInvalidArg("request is nil")
-	}
-	if len(request.GetChunkRequests()) == 0 {
-		return api.NewErrorInvalidArg("no chunk requests provided")
-	}
-	if len(request.GetChunkRequests()) > s.config.MaxKeysPerGetChunksRequest {
-		return api.NewErrorInvalidArg(fmt.Sprintf(
-			"too many chunk requests provided, max is %d", s.config.MaxKeysPerGetChunksRequest))
-	}
-
-	for _, chunkRequest := range request.GetChunkRequests() {
-		if chunkRequest.GetByIndex() == nil && chunkRequest.GetByRange() == nil {
-			return api.NewErrorInvalidArg("chunk request must be either by index or by range")
-		}
-	}
-
-	return nil
-}
-
-// GetChunks retrieves chunks from blobs stored by the relay.
-func (s *Server) GetChunks(ctx context.Context, request *pb.GetChunksRequest) (*pb.GetChunksReply, error) {
-	start := time.Now()
-
-	if s.config.Timeouts.GetChunksTimeout > 0 {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, s.config.Timeouts.GetChunksTimeout)
-		defer cancel()
-	}
-	err := s.validateGetChunksRequest(request)
-	if err != nil {
-		return nil, err
-	}
-
-	s.metrics.ReportChunkKeyCount(len(request.GetChunkRequests()))
-
-	if s.authenticator != nil {
-		client, ok := peer.FromContext(ctx)
-		if !ok {
-			return nil, api.NewErrorInvalidArg("could not get peer information")
-		}
-		clientAddress := client.Addr.String()
-
-		hash, err := s.authenticator.AuthenticateGetChunksRequest(ctx, request)
-		if err != nil {
-			s.metrics.ReportChunkAuthFailure()
-			s.logger.Debug("rejected GetChunks request", "client", clientAddress)
-			return nil, api.NewErrorInvalidArg(fmt.Sprintf("auth failed: %v", err))
-		}
-
-		timestamp := time.Unix(int64(request.GetTimestamp()), 0)
-		err = s.replayGuardian.VerifyRequest(hash, timestamp)
-		if err != nil {
-			s.metrics.ReportChunkAuthFailure()
-			return nil, api.NewErrorInvalidArg(fmt.Sprintf("failed to verify request: %v", err))
-		}
-
-		s.logger.Debug("received authenticated GetChunks request", "client", clientAddress)
-	}
-
-	finishedAuthenticating := time.Now()
-	if s.authenticator != nil {
-		s.metrics.ReportChunkAuthenticationLatency(finishedAuthenticating.Sub(start))
-	}
-
-	clientID := string(request.GetOperatorId())
-	err = s.chunkRateLimiter.BeginGetChunkOperation(time.Now(), clientID)
-	if err != nil {
-		return nil, api.NewErrorResourceExhausted(fmt.Sprintf("rate limit exceeded: %v", err))
-	}
-	defer s.chunkRateLimiter.FinishGetChunkOperation(clientID)
-
-	// keys might contain duplicate keys
-	keys, err := getKeysFromChunkRequest(request)
-	if err != nil {
-		return nil, api.NewErrorInvalidArg(fmt.Sprintf("invalid request: %v", err))
-	}
-
-	mMap, err := s.metadataProvider.GetMetadataForBlobs(ctx, keys)
-	if err != nil {
-		if strings.Contains(err.Error(), blobstore.ErrMetadataNotFound.Error()) {
-			// nolint:wrapcheck
-			return nil, api.NewErrorNotFound(
-				fmt.Sprintf("blob not found, check if blob exists and is assigned to this relay:: %v", keys))
-		}
-		// nolint:wrapcheck
-		return nil, api.NewErrorInternal(fmt.Sprintf("error fetching metadata for blob: %v", err))
-	}
-
-	finishedFetchingMetadata := time.Now()
-	s.metrics.ReportChunkMetadataLatency(finishedFetchingMetadata.Sub(finishedAuthenticating))
-
-	requiredBandwidth, err := computeChunkRequestRequiredBandwidth(request, mMap)
-	if err != nil {
-		return nil, api.NewErrorInternal(fmt.Sprintf("error computing required bandwidth: %v", err))
-	}
-	s.metrics.ReportGetChunksRequestedBandwidthUsage(requiredBandwidth)
-	err = s.chunkRateLimiter.RequestGetChunkBandwidth(time.Now(), clientID, requiredBandwidth)
-	if err != nil {
-		if strings.Contains(err.Error(), "internal error") {
-			return nil, api.NewErrorInternal(err.Error())
-		}
-		return nil, buildInsufficientGetChunksBandwidthError(request, requiredBandwidth, err)
-	}
-	s.metrics.ReportGetChunksBandwidthUsage(requiredBandwidth)
-
-	frames, err := s.chunkProvider.GetFrames(ctx, mMap)
-	if err != nil {
-		return nil, api.NewErrorInternal(fmt.Sprintf("error fetching frames: %v", err))
-	}
-
-	bytesToSend, err := gatherChunkDataToSend(frames, request)
-	if err != nil {
-		return nil, api.NewErrorInternal(fmt.Sprintf("error gathering chunk data: %v", err))
-	}
-
-	s.metrics.ReportChunkDataLatency(time.Since(finishedFetchingMetadata))
-	s.metrics.ReportChunkLatency(time.Since(start))
-
-	return &pb.GetChunksReply{
-		Data: bytesToSend,
-	}, nil
-}
-
-// getKeysFromChunkRequest gathers a slice of blob keys from a GetChunks request.
-func getKeysFromChunkRequest(request *pb.GetChunksRequest) ([]v2.BlobKey, error) {
-	keys := make([]v2.BlobKey, 0, len(request.GetChunkRequests()))
-
-	for _, chunkRequest := range request.GetChunkRequests() {
-		var key v2.BlobKey
-		if chunkRequest.GetByIndex() != nil {
-			var err error
-			key, err = v2.BytesToBlobKey(chunkRequest.GetByIndex().GetBlobKey())
-			if err != nil {
-				return nil, fmt.Errorf("invalid blob key: %w", err)
-			}
-		} else {
-			var err error
-			key, err = v2.BytesToBlobKey(chunkRequest.GetByRange().GetBlobKey())
-			if err != nil {
-				return nil, fmt.Errorf("invalid blob key: %w", err)
-			}
-		}
-		keys = append(keys, key)
-	}
-
-	return keys, nil
-}
-
-// gatherChunkDataToSend takes the chunk data and narrows it down to the data requested in the GetChunks request.
-func gatherChunkDataToSend(
-	frames map[v2.BlobKey]*core.ChunksData,
-	request *pb.GetChunksRequest) ([][]byte, error) {
-
-	bytesToSend := make([][]byte, 0, len(request.GetChunkRequests()))
-
-	for _, chunkRequest := range request.GetChunkRequests() {
-		var framesSubset *core.ChunksData
-		var err error
-
-		if chunkRequest.GetByIndex() != nil {
-			framesSubset, err = selectFrameSubsetByIndex(chunkRequest.GetByIndex(), frames)
-		} else {
-			framesSubset, err = selectFrameSubsetByRange(chunkRequest.GetByRange(), frames)
-		}
-
-		if err != nil {
-			return nil, fmt.Errorf("error selecting frame subset: %v", err)
-		}
-
-		subsetBytes, err := framesSubset.FlattenToBundle()
-		if err != nil {
-			return nil, fmt.Errorf("error serializing frame subset: %v", err)
-		}
-
-		bytesToSend = append(bytesToSend, subsetBytes)
-	}
-
-	return bytesToSend, nil
-}
-
-// selectFrameSubsetByRange selects a subset of frames from a BinaryFrames object based on a range
-func selectFrameSubsetByRange(
-	request *pb.ChunkRequestByRange,
-	allFrames map[v2.BlobKey]*core.ChunksData) (*core.ChunksData, error) {
-
-	key := v2.BlobKey(request.GetBlobKey())
-	startIndex := request.GetStartIndex()
-	endIndex := request.GetEndIndex()
-
-	frames, ok := allFrames[key]
-	if !ok {
-		return nil, fmt.Errorf("frames not found for key %s", key.Hex())
-	}
-
-	if startIndex > endIndex {
-		return nil, fmt.Errorf(
-			"chunk range %d-%d is invalid for key %s, start index must be less than or equal to end index",
-			startIndex, endIndex, key.Hex())
-	}
-	if endIndex > uint32(len(frames.Chunks)) {
-		return nil, fmt.Errorf(
-			"chunk range %d-%d is invald for key %s, chunk count %d",
-			startIndex, endIndex, key, len(frames.Chunks))
-	}
-
-	framesSubset := &core.ChunksData{
-		Chunks:   frames.Chunks[startIndex:endIndex],
-		Format:   frames.Format,
-		ChunkLen: frames.ChunkLen,
-	}
-
-	return framesSubset, nil
-}
-
-// selectFrameSubsetByIndex selects a subset of frames from a BinaryFrames object based on a list of indices
-func selectFrameSubsetByIndex(
-	request *pb.ChunkRequestByIndex,
-	allFrames map[v2.BlobKey]*core.ChunksData) (*core.ChunksData, error) {
-
-	key := v2.BlobKey(request.GetBlobKey())
-	frames, ok := allFrames[key]
-	if !ok {
-		return nil, fmt.Errorf("frames not found for key %s", key.Hex())
-	}
-
-	if len(request.GetChunkIndices()) > len(frames.Chunks) {
-		return nil, fmt.Errorf("too many requested chunks for key %s, chunk count %d",
-			key.Hex(), len(frames.Chunks))
-	}
-
-	framesSubset := &core.ChunksData{
-		Format:   frames.Format,
-		ChunkLen: frames.ChunkLen,
-		Chunks:   make([][]byte, 0, len(request.GetChunkIndices())),
-	}
-
-	for _, index := range request.GetChunkIndices() {
-		if index >= uint32(len(frames.Chunks)) {
-			return nil, fmt.Errorf(
-				"chunk index %d out of range for key %s, chunk count %d",
-				index, key.Hex(), len(frames.Chunks))
-		}
-
-		framesSubset.Chunks = append(framesSubset.Chunks, frames.Chunks[index])
-	}
-
-	return framesSubset, nil
-}
-
-// computeChunkRequestRequiredBandwidth computes the bandwidth required to fulfill a GetChunks request.
-func computeChunkRequestRequiredBandwidth(request *pb.GetChunksRequest, mMap metadataMap) (uint32, error) {
-	requiredBandwidth := uint32(0)
-	for _, req := range request.GetChunkRequests() {
-		var metadata *blobMetadata
-		var key v2.BlobKey
-		var requestedChunks uint32
-
-		if req.GetByIndex() != nil {
-			key = v2.BlobKey(req.GetByIndex().GetBlobKey())
-			metadata = mMap[key]
-			requestedChunks = uint32(len(req.GetByIndex().GetChunkIndices()))
-		} else {
-			key = v2.BlobKey(req.GetByRange().GetBlobKey())
-			metadata = mMap[key]
-
-			if req.GetByRange().GetEndIndex() < req.GetByRange().GetStartIndex() {
-				return 0, fmt.Errorf(
-					"chunk range %d-%d is invalid for key %s, start index must be less than or equal to end index",
-					req.GetByRange().GetStartIndex(), req.GetByRange().GetEndIndex(), key.Hex())
-			}
-
-			requestedChunks = req.GetByRange().GetEndIndex() - req.GetByRange().GetStartIndex()
-		}
-
-		if metadata == nil {
-			return 0, fmt.Errorf("metadata not found for key %s", key.Hex())
-		}
-
-		requiredBandwidth += requestedChunks * metadata.chunkSizeBytes
-	}
-
-	return requiredBandwidth, nil
-}
-
-// buildInsufficientBandwidthError builds an informative error message for when there is insufficient
-// bandwidth to serve a GetChunks() request.
-func buildInsufficientGetChunksBandwidthError(
-	request *pb.GetChunksRequest,
-	requiredBandwidth uint32,
-	originalError error) error {
-
-	chunkCount := 0
-	for _, chunkRequest := range request.GetChunkRequests() {
-		if chunkRequest.GetByIndex() != nil {
-			chunkCount += len(chunkRequest.GetByIndex().GetChunkIndices())
-		} else {
-			chunkCount += int(chunkRequest.GetByRange().GetEndIndex() - chunkRequest.GetByRange().GetStartIndex())
-		}
-	}
-
-	blobCount := len(request.GetChunkRequests())
-
-	return api.NewErrorResourceExhausted(fmt.Sprintf("unable to serve data (%d blobs, %d chunks, %d bytes): %v",
-		blobCount, chunkCount, requiredBandwidth, originalError))
 }
 
 // Start starts the server using the listener provided in the constructor.
